[UPDATE] service details add name category price discount decription and specification. need to add photos,tyoe of events and to anable buttons and comments
