+++ EveAnt/app/src/main/java/com/example/eveant/service/serviceCreate/ServiceCreateFragment3.java
@@ -1,17 +1,16 @@
-package com.example.eveant.serviceEdit;
+package com.example.eveant.service.serviceCreate;
 
 import static android.content.ContentValues.TAG;
 
-import android.annotation.SuppressLint;
 import android.os.Bundle;
+
+import androidx.lifecycle.ViewModelProvider;
+import androidx.navigation.NavController;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.core.content.ContextCompat;
 import androidx.fragment.app.Fragment;
-import androidx.fragment.app.FragmentTransaction;
-import androidx.lifecycle.ViewModelProvider;
-import androidx.navigation.NavController;
 
 import android.util.Log;
 import android.view.LayoutInflater;
@@ -24,7 +23,6 @@ import android.widget.ToggleButton;
 import com.example.eveant.MainActivity;
 import com.example.eveant.R;
 import com.example.eveant.RetrofitClient;
-import com.example.eveant.ServicesViewFragment;
 import com.example.eveant.service.ApiService;
 import com.example.eveant.service.ServiceCreateViewModel;
 import com.example.eveant.service.model.Service;
@@ -35,29 +33,21 @@ import retrofit2.Call;
 import retrofit2.Callback;
 import retrofit2.Response;
 
-
-public class ServiceEditFragment3 extends Fragment {
+public class ServiceCreateFragment3 extends Fragment {
 
     private LinearLayout layoutDuration;
     private LinearLayout layoutRange;
-
-    ServiceCreateViewModel viewModel ;
-
-    @SuppressLint("SetTextI18n")
+    private ServiceCreateViewModel viewModel;
     @Nullable
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.fragment_service_edit3, container, false);
-        viewModel= new ViewModelProvider(requireActivity()).get(ServiceCreateViewModel.class);
-        // Dugme za čuvanje podataka
-        view.findViewById(R.id.save_button).setOnClickListener(v -> {
-            NavController navController = ((MainActivity) getActivity()).getNavController();
-            navController.navigate(R.id.servicesViewFragment);
-        });
+        View view = inflater.inflate(R.layout.fragment_service_create3, container, false);
+
+        viewModel = new ViewModelProvider(requireActivity()).get(ServiceCreateViewModel.class);
 
         view.findViewById(R.id.previous_button).setOnClickListener(v -> {
             NavController navController = ((MainActivity) getActivity()).getNavController();
-            navController.navigate(R.id.serviceEditFragment2);
+            navController.navigate(R.id.serviceCreateFragment2);
         });
 
         ToggleButton manualButton = view.findViewById(R.id.manualButton);
@@ -71,6 +61,8 @@ public class ServiceEditFragment3 extends Fragment {
             }
         });
 
+
+
         automaticButton.setOnCheckedChangeListener((buttonView, isChecked) -> {
             if (isChecked) {
                 manualButton.setChecked(false);
@@ -79,11 +71,9 @@ public class ServiceEditFragment3 extends Fragment {
             }
         });
 
-        // Inicijalizacija layout-ova
         layoutDuration = view.findViewById(R.id.layoutDuration);
         layoutRange = view.findViewById(R.id.layoutRange);
 
-        // Postavljanje klik listener-a
         layoutDuration.setOnClickListener(v -> {
             resetSelections();
             layoutDuration.setBackgroundResource(R.drawable.background_selected);
@@ -94,71 +84,19 @@ public class ServiceEditFragment3 extends Fragment {
             layoutRange.setBackgroundResource(R.drawable.background_selected);
         });
 
+
         EditText reservationPeriod = view.findViewById(R.id.deadline);
         EditText cancellationPeriod = view.findViewById(R.id.cacnellationPeriod);
-        EditText durationHoursField = view.findViewById(R.id.durationHours);
-        EditText durationMinutesField = view.findViewById(R.id.durationMinutes);
-        EditText minHours = view.findViewById(R.id.minHours);
-        EditText minMinutes = view.findViewById(R.id.minMinutes);
-        EditText maxHours = view.findViewById(R.id.maxHours);
-        EditText maxMinutes = view.findViewById(R.id.maxMinutes);
-
-        Service service = viewModel.getService().getValue();
-        Log.d(TAG, "onCreateView: "+service.getName());
-
-        if (service != null) {
-            Log.d(TAG, "onCreateView: ovde nastaje bagggggggggggg");
-            Integer minDuration = service.getMinEngagement() != null ? service.getMinEngagement() : 0;
-            Integer maxDuration = service.getMaxEngagement() != null ? service.getMaxEngagement() : 0;
-
-            if (maxDuration == 0) {
-                // Sve ide u duration
-                int hours = minDuration / 60;
-                int minutes = minDuration % 60;
-
-                durationHoursField.setText(String.valueOf(hours));
-                durationMinutesField.setText(String.valueOf(minutes));
-                layoutDuration.setBackgroundResource(R.drawable.background_selected);
-            } else {
-                // Postoje min i max duration
-                int minHoursValue = minDuration / 60;
-                int minMinutesValue = minDuration % 60;
-
-                int maxHoursValue = maxDuration / 60;
-                int maxMinutesValue = maxDuration % 60;
-
-                minHours.setText(String.valueOf(minHoursValue));
-                minMinutes.setText(String.valueOf(minMinutesValue));
-                maxHours.setText(String.valueOf(maxHoursValue));
-                maxMinutes.setText(String.valueOf(maxMinutesValue));
-                layoutRange.setBackgroundResource(R.drawable.background_selected);
-            }
-        } else {
-            Log.e(TAG, "Service object is null");
-        }
-
-        Log.d(TAG, "onCreateView: evo meeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee"+service.getAutomation());
-
-        reservationPeriod.setText(service.getReservationDeadLine().toString());
-        cancellationPeriod.setText(service.getCancellationPeriod().toString());
-
-        if(service.getAutomation()==null){
-
-        }else if(!service.getAutomation()){
-            manualButton.setChecked(true);
-        }else{
-            automaticButton.setChecked(true);
-        }
-
-
 
 
         view.findViewById(R.id.save_button).setOnClickListener(v -> {
+            Service service = viewModel.getService().getValue();
             if (service != null) {
                 if (layoutDuration.getBackground().getConstantState().equals(
                         ContextCompat.getDrawable(requireContext(), R.drawable.background_selected).getConstantState())) {
 
-
+                    EditText durationHoursField = view.findViewById(R.id.durationHours);
+                    EditText durationMinutesField = view.findViewById(R.id.durationMinutes);
 
                     String strHours = durationHoursField.getText().toString();
                     String strMinutes = durationMinutesField.getText().toString();
@@ -172,6 +110,10 @@ public class ServiceEditFragment3 extends Fragment {
                 } else if (layoutRange.getBackground().getConstantState().equals(
                         ContextCompat.getDrawable(requireContext(), R.drawable.background_selected).getConstantState())) {
 
+                    EditText minHours = view.findViewById(R.id.minHours);
+                    EditText minMinutes = view.findViewById(R.id.minMinutes);
+                    EditText maxHours = view.findViewById(R.id.maxHours);
+                    EditText maxMinutes = view.findViewById(R.id.maxMinutes);
 
                     String strHours = minHours.getText().toString();
                     String strMinutes = minMinutes.getText().toString();
@@ -202,7 +144,7 @@ public class ServiceEditFragment3 extends Fragment {
                 service.setAutomation(automaticButton.isChecked());
 
                 viewModel.updateService(service);
-                updateService();
+                saveService();
             }
         });
 
@@ -210,36 +152,35 @@ public class ServiceEditFragment3 extends Fragment {
         return view;
     }
 
-    // Ispravno definisana metoda za resetovanje selekcije
     private void resetSelections() {
         layoutDuration.setBackgroundResource(R.drawable.background_unselected);
         layoutRange.setBackgroundResource(R.drawable.background_unselected);
     }
 
-    private void updateService() {
+    private void saveService() {
         Service serviceToSave = viewModel.getService().getValue();
 
         ServiceDTO serviceDTO = ServiceMapper.INSTANCE.toDTO(serviceToSave);
 
-        Log.d(TAG, "updateService:ovo gledaj to mi trebaaaaa "+serviceDTO);
-        if (serviceToSave != null) {
+        Log.d(TAG, "saveService: "+serviceDTO.getPrice());
+        if (serviceDTO != null) {
             ApiService apiService = RetrofitClient.apiService;
-            Log.d("ServiceToSave", "Service to save: " + serviceToSave.toString());
+            Log.d("ServiceToSave", "Service to save: " + serviceDTO.toString());
 
-            apiService.updateService(serviceToSave.getId(),serviceDTO).enqueue(new Callback<Void>() {
+            apiService.createService(serviceDTO).enqueue(new Callback<Void>() {
                 @Override
                 public void onResponse(Call<Void> call, Response<Void> response) {
                     if (response.isSuccessful()) {
                         NavController navController = ((MainActivity) getActivity()).getNavController();
-                        navController.navigate(R.id.servicesViewFragment);
+                        navController.navigate(R.id.actionCreateFragment_toViewServices);
                     } else {
-                        Log.e("Update servie", "Failed to save service. Response code: " + response.code());
+                        Log.e("SaveService", "Failed to save service. Response code: " + response.code());
                     }
                 }
 
                 @Override
                 public void onFailure(Call<Void> call, Throwable t) {
-                    Log.e("UpdateService", "Error occurred: " + t.getMessage());
+                    Log.e("SaveService", "Error occurred: " + t.getMessage());
                 }
             });
         }
