+++ EveAnt/app/src/main/java/com/example/eveant/service/serviceEdit/ServiceEditFragment2.java
@@ -1,4 +1,4 @@
-package com.example.eveant.serviceEdit;
+package com.example.eveant.service.serviceEdit;
 
 import android.os.Bundle;
