+++ EveAnt/app/src/main/res/layout/fragment_service_details.xml
@@ -35,7 +35,7 @@
                     android:fontFamily="@font/poppins_bold"
                     android:text="Service name"
                     android:textAppearance="@style/place"
-                    android:textSize="20sp"
+                    android:textSize="25sp"
                     android:textStyle="bold" />
 
 
@@ -43,7 +43,6 @@
             <HorizontalScrollView
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:layout_marginTop="16dp"
                 android:padding="8dp"
                 android:scrollbars="none">
 
@@ -54,8 +53,8 @@
 
                     <!-- Slika 1 -->
                     <ImageView
-                        android:layout_width="200dp"
-                        android:layout_height="150dp"
+                        android:layout_width="400dp"
+                        android:layout_height="250dp"
                         android:layout_marginEnd="8dp"
                         android:scaleType="centerCrop"
                         android:src="@drawable/pastry"
@@ -63,8 +62,8 @@
 
                     <!-- Slika 2 -->
                     <ImageView
-                        android:layout_width="200dp"
-                        android:layout_height="150dp"
+                        android:layout_width="400dp"
+                        android:layout_height="250dp"
                         android:layout_marginEnd="8dp"
                         android:scaleType="centerCrop"
                         android:src="@drawable/pastry"
@@ -72,8 +71,8 @@
 
                     <!-- Slika 3 -->
                     <ImageView
-                        android:layout_width="200dp"
-                        android:layout_height="150dp"
+                        android:layout_width="400dp"
+                        android:layout_height="250dp"
                         android:scaleType="centerCrop"
                         android:src="@drawable/pastry"
                         android:background="@drawable/rounded_corners_white" />
@@ -178,7 +177,7 @@
                 android:fontFamily="@font/poppins_regular" />
 
             <TextView
-                android:id="@+id/specifics"
+                android:id="@+id/specification"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:layout_margin="16dp"
@@ -200,18 +199,18 @@
                 <RelativeLayout
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
-                    android:layout_marginEnd="8dp">
+                    android:layout_marginEnd="90dp">
                     <!-- Nova cena -->
                     <TextView
                         android:id="@+id/newPrice"
                         android:layout_width="wrap_content"
                         android:layout_height="wrap_content"
-                        android:layout_marginTop="30dp"
-                        android:layout_marginRight="90dp"
+                        android:layout_marginTop="55dp"
+                        android:layout_marginRight="150dp"
                         android:layout_gravity="end"
                         android:text="8$/hr"
                         android:textColor="@color/light_purple"
-                        android:textSize="20sp"
+                        android:textSize="25sp"
                         android:fontFamily="@font/poppins_bold"
                         android:layout_alignParentEnd="true"/>
 
@@ -228,26 +227,26 @@
 
                         <TextView
                             android:id="@+id/oldPrice"
-                            android:layout_width="60dp"
-                            android:layout_height="50dp"
+                            android:layout_width="120dp"
+                            android:layout_height="100dp"
                             android:text="10$/hr"
                             android:textColor="#000000"
-                            android:textSize="18sp"
+                            android:textSize="25sp"
                             android:fontFamily="@font/poppins_regular"
                             android:gravity="center" />
                     </LinearLayout>
 
                     <TextView
                         android:id="@+id/discountBadge"
-                        android:layout_width="40dp"
-                        android:layout_height="30dp"
+                        android:layout_width="60dp"
+                        android:layout_height="50dp"
                         android:layout_alignParentTop="true"
                         android:layout_alignParentEnd="true"
                         android:background="@drawable/rounded_rectangle_purple"
                         android:padding="4dp"
                         android:text="10%"
                         android:textColor="@android:color/white"
-                        android:textSize="14sp"
+                        android:textSize="20sp"
                         android:textStyle="bold"
                         android:fontFamily="@font/poppins_regular"
                         android:gravity="center" />
@@ -265,18 +264,7 @@
                 android:layout_marginTop="16dp"
                 android:padding="16dp">
 
-                <!-- Ljubičasto dugme - View Comments & Reviews -->
-                <Button
-                    android:id="@+id/btn_view_reviews"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:layout_marginBottom="8dp"
-                    android:background="@drawable/rounded_corners_light_blue"
-                    android:text="@string/view_comments_and_reviews"
-                    android:textColor="@android:color/white"
-                    android:textSize="14sp"
-                    android:fontFamily="@font/poppins_medium"
-                    android:textStyle="bold" />
+
 
                 <!-- Ljubičasto dugme - Buy Product -->
                 <Button
